-Created Logic for keys.equip to lock/unlock a door(80%)
I had to figure out how I'm going to get the player Pos to do a scene trace. I had to end up creating/setting a game object ref for the Player controller Game object.

So now any equipment will be able to fire with anything needed from the player Controller.

Still need to add a check to see if the player owns the door or not.