-Created Jail Timer UI/Created Extra Unarrested Logic
I got the Jail timer to show up/hide with the total time and it even ticks down!

I also wanted to have the Unarrest System to properly check for Sweps that are 'Starting Sweps' or 'Legal Sweps'. If so, don't remove the weapon.

I also added another setting to the DarkRP Gamemode to allow the Owners to Allow/Disallow Illegal Sweps  being taken away when unarrested. Inorder to do this I had to go in and change some of the Vars in the DarkRP Gamemode to be lists instead of arrays. I might change this back if there is any problems.