-Further Tried to fix/Create Logic for Unarresting players
The specific thing is adding the correct weapons. So I had to go in and add extra variable to the Swep Hotbar system to allow me to check if a swep is okay for the player to have when they are unarrested.

THIS is super crazy for me because using the For Loops in S&Box compared to UE5 is CRAZY!!.. I'm learning though and figured out what I needed for the logic to begin working and checking the sweps. Now I need to fix like 2 bugs and the system should work.